Cleaned up controllers and Dom, added sticky footer, cleaned unecessary ajax calls on search
