+++ public/css/wiifat.css
@@ -17,10 +17,28 @@ body {
   height: 207px;
 }
 
-#goToChat {
-  position: fixed;
-  bottom: 0;
-  right: 15px;
+/* Helpers */
+.compressed {
+  max-height: 0;
+  visibility: hidden;
+  opacity: 0;
+  padding: 0 !important;
+  -webkit-transition: height 500ms ease-in-out, opacity 250ms ease-in-out, visibility 250ms ease-in-out;
+  -moz-transition: height 500ms ease-in-out, opacity 250ms ease-in-out, visibility 250ms ease-in-out;
+  -ms-transition: height 500ms ease-in-out, opacity 250ms ease-in-out, visibility 250ms ease-in-out;
+  -o-transition: height 500ms ease-in-out, opacity 250ms ease-in-out, visibility 250ms ease-in-out;
+  transition: height 500ms ease-in-out, opacity 250ms ease-in-out, visibility 250ms ease-in-out;
+}
+
+.expanded {
+  max-height: 100%;
+  visibility: visible;
+  opacity: 1;
+  -webkit-transition-delay: 0ms;
+  -moz-transition-delay: 0ms;
+  -ms-transition-delay: 0ms;
+  -o-transition-delay: 0ms;
+  transition-delay: 0ms;
 }
 
 /* Team page */
@@ -63,41 +81,71 @@ body {
 }
 
 /* Chat page */
-#chat {
-  width: 400px;
-  max-width: 80%;
-  display: block;
-  margin: auto;
-  border: 2px solid #000;
-  border-bottom: none;
-  border-top-left-radius: 25px;
-  border-top-right-radius: 25px;
-  padding: 10px;
+#goToChat,
+.go-to-chat {
   position: fixed;
   bottom: 0;
-  right: 96px;
-  background: #e9ecef;
 }
 
-#chat .pages {
-  list-style: none;
-  padding: 0;
+#chatWrapper {
+  position: fixed;
+  z-index: 10000;
+  text-align: center;
+  height: 500px;
+  width: 400px;
+  max-width: 100%;
+  right: 0;
+  left: 0;
+  bottom: -458px;
 }
 
-#chat .pages .chatArea {
+#chatWrapper .nav-tabs {
+  border-bottom: 1px solid #e9ecef;
+}
+
+#chatWrapper .nav-tabs #letsChat {
+  background-color: #e9ecef;
+  border-top: 1px solid #000;
+  border-left: 1px solid #000;
+  border-right: 1px solid #000;
+  border-bottom: 1px solid #e9ecef;
+  margin-bottom: -2px;
+}
+
+#chatWrapper.slide-up {
+  bottom: 0px !important;
+}
+
+#chatWrapper.slide-down {
+  bottom: -458px !important;
+}
+
+#chatWrapper #chat {
+  width: 400px;
+  max-width: 100%;
   height: 100%;
-  padding-bottom: 60px;
-  padding-left: 0;
+  display: block;
+  padding: 10px;
+  background: #e9ecef;
+  border-top: 1px solid #000;
+  border-left: 1px solid #000;
+  border-right: 1px solid #000;
+  border-top-right-radius: .25rem;
 }
 
-#chat .pages .chatArea .username {
+#chatWrapper #chat .chatArea {
+  height: 80%;
+  overflow-y: scroll;
+}
+
+#chatWrapper #chat .chatArea .username {
   font-weight: 700;
   overflow: hidden;
   padding-right: 15px;
   text-align: right;
 }
 
-#chat .pages .chatArea .messages {
+#chatWrapper #chat .chatArea .messages {
   height: 100%;
   margin: 0;
   overflow-y: scroll;
@@ -106,13 +154,29 @@ body {
   padding: 0;
 }
 
-#chat .pages .chatArea .log {
+#chatWrapper #chat .chatArea .log {
   color: gray;
   font-size: 70%;
   margin: 5px;
   list-style: none;
 }
 
-#chat .pages .chatArea .message.typing .messageBody {
+#chatWrapper #chat .chatArea .message.typing .messageBody {
   color: gray;
+}
+
+#chatWrapper #chat #chatForm {
+  position: absolute;
+  right: 0;
+  left: 0;
+  bottom: 0;
+}
+
+@media screen and (max-width: 480px) {
+  #chatWrapper {
+    width: 100%;
+  }
+  #chatWrapper #chat {
+    width: 100%;
+  }
 }
