+++ public/js/search.js
@@ -5,40 +5,55 @@ $(() => {
         let searchInput = $('#searchBar').val().trim();
         let searchBy = $('#searchBy').val();
         if (searchBy === "title") {
-            $.ajax('/api/books/title/' + searchInput, {
-                type: "GET",
-            }).then(result => {
-                console.log(result);
-                window.location.href = "/api/books/title/" + searchInput;
-            });
+            window.location.href = "/search/books/title/" + searchInput;
         } else if (searchBy === "genre") {
-            $.ajax('/api/books/genre/' + searchInput, {
-                type: "GET",
-            }).then(result => {
-                console.log(result);
-                window.location.href = "/api/books/genre/" + searchInput;
-            });
+            window.location.href = "/search/books/genre/" + searchInput;
         } else if (searchBy === "author") {
-            $.ajax('/api/books/author/' + searchInput, {
-                type: "GET",
-            }).then(result => {
-                console.log(result);
-                window.location.href = "/api/books/author/" + searchInput;
-            });
+            window.location.href = "/search/books/author/" + searchInput;
         } else if (searchBy === "book-id") {
-            $.ajax('/api/books/id/' + searchInput, {
-                type: "GET",
-            }).then(result => {
-                console.log(result);
-                window.location.href = "/api/books/id/" + searchInput;
-            });
+            window.location.href = "/search/books/id/" + searchInput;
         } else if (searchBy === "author-id") {
-            $.ajax('/api/books/author/' + searchInput, {
-                type: "GET",
-            }).then(result => {
-                console.log(result);
-                window.location.href = "/api/books/author/" + searchInput;
-            });
+            window.location.href = "/search/books/author/" + searchInput;
         }
-    })
+    });
+    // searchBtn.click(() => {
+    //     let searchInput = $('#searchBar').val().trim();
+    //     let searchBy = $('#searchBy').val();
+    //     if (searchBy === "title") {
+    //         $.ajax('/search/books/title/' + searchInput, {
+    //             type: "GET",
+    //         }).then(result => {
+    //             console.log(result);
+    //             window.location.href = "/search/books/title/" + searchInput;
+    //         });
+    //     } else if (searchBy === "genre") {
+    //         $.ajax('/search/books/genre/' + searchInput, {
+    //             type: "GET",
+    //         }).then(result => {
+    //             console.log(result);
+    //             window.location.href = "/search/books/genre/" + searchInput;
+    //         });
+    //     } else if (searchBy === "author") {
+    //         $.ajax('/search/books/author/' + searchInput, {
+    //             type: "GET",
+    //         }).then(result => {
+    //             console.log(result);
+    //             window.location.href = "/search/books/author/" + searchInput;
+    //         });
+    //     } else if (searchBy === "book-id") {
+    //         $.ajax('/search/books/id/' + searchInput, {
+    //             type: "GET",
+    //         }).then(result => {
+    //             console.log(result);
+    //             window.location.href = "/search/books/id/" + searchInput;
+    //         });
+    //     } else if (searchBy === "author-id") {
+    //         $.ajax('/search/books/author/' + searchInput, {
+    //             type: "GET",
+    //         }).then(result => {
+    //             console.log(result);
+    //             window.location.href = "/search/books/author/" + searchInput;
+    //         });
+    //     }
+    // });
 });
