+++ routes/apiRoutes.js
@@ -10,15 +10,6 @@ module.exports = function(app, passport) {
 
     app.delete('/api/book/:id', apiController.deleteBook);
 
-    // search
-    app.get('/api/books/id/:id', apiController.searchBooksById);
-
-    app.get('/api/books/title/:title', apiController.searchBooksByTitle);
-
-    app.get('/api/books/author/:id', apiController.searchBooksByAuthor);
-
-    app.get('/api/books/genre/:genre', apiController.searchGenre);
-
     // posts
     app.post('/api/book/post', apiController.addPost);
