+++ views/book.handlebars
@@ -1,5 +1,3 @@
-{{> header}}
-
 <div class="container" id="editBook">
     <div class="container-fluid">
         <h1 class="display-4">{{book.title}}</h1>
@@ -58,5 +56,3 @@
         </div>
     </div>
 </div>
-
-{{> footer}}
