+++ views/dashboard.handlebars
@@ -1,28 +1,9 @@
-<!-------------------------------------
-+++++ Delete Wwhen Done +++++
-user.firstname
-user.lastname
-user.username
-user.about
-user.photo
-user.email
---------------------------------------->
-<!-------------------------------------
-+++++ Delete Wwhen Done +++++
-books
-this.Book.id
-this.Book.title
-this.Book.body
-this.Book.genre
-this.Book.imageUrl
---------------------------------------->
-{{> header}}
 <div class="container">
     <div class="jumbotron">
         <div class="container">
             <div class="row">
                 <div class="col-sm-12">
-                    <i class="fas fa-cog float-right" data-toggle="modal" data-target="#exampleModal"></i>
+                    <i class="fas fa-cog float-right" data-toggle="modal" data-target="#updateUserModal"></i>
                 </div>
             </div>
             <div class=" row">
@@ -49,7 +30,8 @@ this.Book.imageUrl
                             <div class="col-md-3 mb-3">
                                 <div class="card book">
                                     <a href="/book/{{this.Book.id}}">
-                                        <img src="{{this.Book.imageUrl}}" alt="{{this.Book.title}}" class="card-img w-100 d-block">
+                                        <img src="{{this.Book.imageUrl}}" alt="{{this.Book.title}}"
+                                            class="card-img w-100 d-block">
                                         <div class="card-img-overlay">
                                             <h4>{{this.Book.title}}</h4>
                                         </div>
@@ -64,9 +46,7 @@ this.Book.imageUrl
         </div>
     </div>
 </div>
-
-</div>
-<div class="modal fade" id="exampleModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel"
+<div class="modal fade" id="updateUserModal" tabindex="-1" role="dialog" aria-labelledby="updateUserModalLabel"
     aria-hidden="true">
     <div class="modal-dialog" role="document">
         <div class="modal-content">
@@ -92,6 +72,4 @@ this.Book.imageUrl
         </div>
     </div>
 </div>
-
-{{> footer}}
 {{> dashboard-scripts}}
