+++ views/homepage.handlebars
@@ -1,4 +1,3 @@
-{{> header}}
 <div class="article-list">
     <div class="jumbotron" id="homeBanner">
         <div class="intro my-5">
@@ -108,7 +107,4 @@
         </div>
     </div>
 </div>
-
-{{> footer}}
-
 {{> login-modal}}
