+++ views/login.handlebars
@@ -1,4 +1,3 @@
-{{> header}}
 <!-- Start: Registration Form with Photo -->
 <div class="register-photo">
     <!-- Start: Form Container -->
@@ -44,6 +43,4 @@
     <!-- End: Form Container -->
 </div>
 <!-- End: Registration Form with Photo -->
-{{> footer}}
-
 {{> login-modal}}
