+++ views/partials/chat-scripts.handlebars
@@ -1,5 +1,25 @@
 <script>
 $(function() {
+
+    {{!-- $('#letsChat').on('click', function(e) {
+        e.preventDefault();
+        $chat = $('#chat');
+        $chat.toggleClass('expanded');
+    }); --}}
+
+      $('#letsChat').click(function (e) {
+          console.log('clicked');
+        e.preventDefault();
+        var $chat = $('#chatWrapper');
+        if ($chat.hasClass('slide-up')) {
+            $chat.addClass('slide-down', 1000, 'easeOutBounce');
+            $chat.removeClass('slide-up');
+        } else {
+            $chat.removeClass('slide-down');
+            $chat.addClass('slide-up', 1000, 'easeOutBounce');
+        }
+    });
+
     var FADE_TIME = 150; // ms
     var TYPING_TIMER_LENGTH = 400; // ms
     var COLORS = [
@@ -218,7 +238,7 @@ $(function() {
     socket.on('login', (data) => {
         connected = true;
         // Display the welcome message
-        var message = "Welcome";
+        var message = "Welcome, " + username;
         log(message, {
             prepend: true
         });
