+++ views/result.handlebars
@@ -1,5 +1,3 @@
-{{> header}}
-
 <div class="container">
     <h1>Results</h1>
     <hr>
@@ -27,7 +25,4 @@
     {{/each}}
     {{/if}}
 </div>
-
-{{> footer}}
-
 {{> login-modal}}
